Abstract finish class to own file
